Fix permissions when copying from ObjectStorage
Make sure that when a user copy a file from a directory they don't have
many permissions to a directory there they have permissions, the
permissions are correctly set.

This fix a bug there copying a file from a groupfolder there an user has
only read permissions to their home folder, they can then delete the
file.

I'm not completely sure this is the right way since we should check that
permissions the target folder has. Maybe the user is not supposed to
delete it after adding the file (e.g. share folder with deletion
forbidden).

Signed-off-by: Carl Schwan <carl@carlschwan.eu>